fix: Fix object angle conversion, add tests, bump dependencies (#9)
* fix: Set object angle type to float64

* chore: Bump to go v1.20

* chore: Add test TestConvertExcalidrawToGliffy

* chore: Add GHA CI worklow

* chore: Configure Renovate to use semantic commits

* chore: Bump goreleaser-action to v4

* chore: Add more complex test data

* chore: Update readme

---------

+++ internal/datastructures/excalidraw.go
@@ -6,7 +6,7 @@ type ExcalidrawScene struct {
 		ViewBackgroundColor string      `json:"viewBackgroundColor"`
 	} `json:"appState"`
 	Elements []struct {
-		Angle           int64         `json:"angle"`
+		Angle           float64       `json:"angle"`
 		BackgroundColor string        `json:"backgroundColor"`
 		Baseline        float64       `json:"baseline"`
 		BoundElementIds []string      `json:"boundElementIds"`
